mjd@alphalink.com.au: "Build Railroad" never shows up in the Orders menu, even if it is a valid option.  I have rigged this so that "Build Railroad" replaces "Build Road" in the Options menu when appropriate.
As I was working on this fix, I noticed a bug:  If you click on a unit
which is engaged in some activity (pollution cleanup, road building, etc.)
the Orders menu does not reflect the fact that this unit is idle and can
accept new orders until you click on that unit a SECOND time. I found a fix
for this problem and incorporated it into this patch as well.
Patch submitted by Cedric Tefft <cedric@earthling.net>.


git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@553 a0f10bec-cc02-0410-94fc-a9cfff90b4cd