mjd@alphalink.com.au: Some changes so that a score file is generated which can be used to generate graphs of different statistics over time. This is very similar to the graphs generated in CivI at the end of the game.
When the "scorelog" server option is turned on, a file called
"civscore.log" is created and updated every turn.  A post-
processing perl script (which can be found in the Freeciv contrib
dir) is then used to parse the log file and generate the output
files, which are formatted for "xgraph".

This was submitted by Trevor Pering <pering@tongass.EECS.Berkeley.EDU>.


git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@577 a0f10bec-cc02-0410-94fc-a9cfff90b4cd