Fix for bug PR#27 (the city dialog icons not refreshing properly), submitted by Vasco Alexandre Da Silva Costa.
Basically what this does is hide those widgets and then show them again
(to force a redraw).

There's probably a better way to do this by changing the code for pixmap
widgets handling (which probably has a bug which causes this), but that
would take way too much time.  This little hack should to the trick.


git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@1412 a0f10bec-cc02-0410-94fc-a9cfff90b4cd