Adds display of production values to main map "city tiles" display.
Also, fixes bug where if a city was on one side of the zero/maximum-x
divide, and the tile you shift-clicked on was on the other side, then
nothing would happen.  This was because adjust_workers() did its own
conversion from map to city coordinates, rather than using the
map_to_city_[xy]() routines, and this conversion code was flawed.


git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@1891 a0f10bec-cc02-0410-94fc-a9cfff90b4cd