Removed the bug/feature whereby a city with zero shield surplus was given a "free" shield each turn.
As a related move, I also changed the "if" condition used to check for
the need to disband units due to lack of shields for upkeep to a "while"
loop, which means as many units will be disbanded as necessary in order to
restore the shield surplus of the city to at least zero.  Previously, only
one unit would be disbanded in this situation, and you get a free shield
surplus even if it's actually still a deficit.

Patch submitted by Mike Jing <miky40@hotmail.com>.


git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@1996 a0f10bec-cc02-0410-94fc-a9cfff90b4cd