Fixes a few small problems with city and unit visibility.  It also does a little clean-up.
The city visibility problem is that when the focus unit is in a city, and
its icon is flashing, when the unit icon was seen, _nothing_ of the city
was -and- when the unit icon was not seen, the city's "occupied" flag was
not shown.

Now, the city is always fully displayed "under" the flashing unit icon.

The unit visibility problem is one that was introduced since 1.10.0
(though, I don't know exactly when).  When a unit is defending a city, the
attacker should see the unit.  Currently, the attacker only sees the city.

Now, the defending unit is visible to the attacker.

Fixed player_color to return a color in proper range (broken when
MAX_NUM_PLAYERS was increased).


git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@2362 a0f10bec-cc02-0410-94fc-a9cfff90b4cd