This patch changes the semantics of the 'cmdlevel first' feature. The command
/cmdlevel ctrl first

allows 'ctrl' level to the first connection that takes it.

If no connections exist yet this will be the first connection made.
If connections already exist, or in the case the controling player
disconnects the situation changes with this patch:
new connections will no longer obtain this level, but it must be
grabbed explicitly by a connected player, using the new command

/firstlevel

The patch also corrects some minor inconsistencies in interpunction etc.

Patch by Reinier Post <rp@win.tue.nl>


git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@2872 a0f10bec-cc02-0410-94fc-a9cfff90b4cd