Waypoints and goto line drawing cleanup.
When specifying a goto route in the client you can now hit "g" again to set
a waypoint that the unit will go to first. Similarly, once in the patrol
state you can set a waypoint by hitting the patrol shortcut key again.
The line drawing has been recoded, and we now draw a cyan line for the
route. Moving the cursor will now result in as few redraws as possible.

Also fixes bugs:
-The goto destination line is now automatically updated when you recenter
 the screen.
-fix patrol bug when route to destination was not found.
-fix bug that ships could not target a land unit with the client goto.
-The line drawing code would in rare cases leave a pixel when undrawing.
-fix client goto bug where ships wouldn't go from unknown terain onto ocean.


git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@3115 a0f10bec-cc02-0410-94fc-a9cfff90b4cd