[quote] The new gettext uses information from the header of each PO file to convert the text of MO files to whatever is required on the fly.  For instance, this would allow one to use a no_NO.ISO-8859-1 MO file in a no_NO.UTF-8 locale.
Unfortunately, this information is missing or incorrect for most of
the PO files supplied with Freeciv, so I went through and
added/corrected this information as necessary.  I also tidied them up
by replacing e.g. "latin1-iso8859-1" with "ISO-8859-1", which the
gettext manual says is necessary.

Note that I also changed the en_GB entry from "7bit" to "8bit".  This
may seem odd, but the manual recommends it and it is unlikely to cause
any harm.

The header entries are currently marked as fuzzy, so that this
information will not be available in any case.  However, that is an
issue that can and should be dealt with seperately.
[/quote]

Patch by Gaute B Strokkenes <gs234@cam.ac.uk>


git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@3404 a0f10bec-cc02-0410-94fc-a9cfff90b4cd