Remove the "focus hack". The purpose of this "hack" is unknown. This "hack" introduce a non-normal unit position which obstructs the task to bring normal positions to all code.
Quoting from an email from Jason Dorje Short <vze2zq63@verizon.net>:
-----------
The focus hack changes the position of the unit so that it is not
drawn in refresh_tile_mapcanvas below.  (The unit is in focus, of
course, since it's being moved.)  Later in refresh_tile_mapcanvas, a
check is done to see if the unit in focus is on the current square,
and if so it's drawn.  Since the unit has been moved, it will not be
found on the current square but since the coordinates of the unit are
invalid a call to map_get_tile for those coordinates is not "clean".
Under alternate topologies the behavior may become unpredictable.

It does not appear as though this hack is needed; it is very old and I
see no problems with drawing when it is removed.  If it is to be kept,
there are several ways it could be handled.  An is_normal_map_pos
check could be done before the map_get_tile call, or the unit could be
removed from focus entirely (which would solve the problem but may
have side effects).
-----------


git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@3892 a0f10bec-cc02-0410-94fc-a9cfff90b4cd