Correct the condition for which the techs aren't updated to show the techs of a disconnected player during the main game.
The problem: init_tech (which marks all techs given in the ruleset
unknown, makes A_NONE and gives init techs) is only called in a new
game and if after all players chose their nations. So this leaves the
time window from the time a player connects (and this is published to
all other players) till the time this player disconnects before the
init_tech is called. In this time window the tech data is unset. This
is the problem that the previous patch should have fixed. However the
previous patch was too restrictive.

Reported by Christian Knoke <chrisk@enter.de>. Patch from me.


git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@6049 a0f10bec-cc02-0410-94fc-a9cfff90b4cd