This patch cleans the usage of pcity->occupied up in two ways:
1.  Instead of pcity->occupied it's now pcity->client.occupied.  This
makes it harder to accidentally use it at the server (which was one of
the original bugs with this code), and is a prelude for more
client-side-only fields (see PR#3771).  Of course this means all users
have to be changed.

2.  Lots of comments added about the client-side usage of this field.
It seems that the client side of the code is correct, so some explanations
are put into the comments.

Patch by Jason.


git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@6546 a0f10bec-cc02-0410-94fc-a9cfff90b4cd