Fix some bugs and reorder things slightly in ai_gothere.
Reordering:
Before, the unit would try to find_ferry whenever it wanted ferry, but
then if the ferry is not nearby, the result would not be used
constructively.
Now the unit will use find_ferry only if it's totally lost (inland), then
it'd walk towards the coast.  If unit is on the coast it will wait for a
ferry to come within one step.

Bugs were:
1. Ferry not made available when released.
2. Ferry not using omniscience when looking for psngrs.
3. Units reserving a ferry alhtough it does them no good and it hurts
other things.
4. Units reserving other players' ferries.

Issued as (PR#8827).


git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@7684 a0f10bec-cc02-0410-94fc-a9cfff90b4cd