This patch adds a new function get_current_construction_bonus.
The function does the same thing it does in the effects patch: it takes
a city and an effect, and returns the effect bonus (aka "power")
provided by the current production.

It doesn't access the effect arrays (which are incomplete). Instead
it's just a wrapper for checking the improvement enum for the current
production. Also it's only used in a few simple places that previously
checked B_CAPITAL - only EFT_PROD_TO_GOLD is supported.

Patch by Jason in PR#9815.


git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@8064 a0f10bec-cc02-0410-94fc-a9cfff90b4cd