Rewriting larger portions of the AI diplomacy code, in order to get rid of the 'lobster logic' (always attack the stronger player) approach. Instead, the AI becomes opportunistic and greedy on players with lots of cities and little defense, and is influenced 'emotionally' by players who are nice to it throughout the game.
The previous concept of countdown torwards a single war target has been replaced by a
more general 'war footing' concept, which allows countdowns to war against multiple
players at once. When on war footing (preparing for an attack), the AI will give
priority to taxes (gold) and military.

It should be slower to get treaties with the AI now, but you can get peace if you really
want, as opposed to the 'if you are the lobster, nothing helps' of the current code.

The AI will now inform allies when attacked or preparing for war.

Logging has been improved considerably.

See PR#13524.


git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@10819 a0f10bec-cc02-0410-94fc-a9cfff90b4cd