fix creation of tropical temperature range
For low values of the temperature setting TROPICAL_LEVEL evaluates
to MAX_COLATITUDE. Due to the fact, that the temperature_map did _not_
reach this value at the equator, tropical climate was set next to the
equator but temperate climate at the equator.

solution:

set 90% of the MAX_COLATITUDE as one of the values within the definition
of TROPICAL_LEVEL.

Additionally fix the comment above the definitions.

See gna bug #15310



git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@16810 a0f10bec-cc02-0410-94fc-a9cfff90b4cd