Make impossible to access operating system from Lua scripts
For security reasons, Lua scripts should not be able to read files or
run programs on the host computer; freeciv scenarios should only be
able to influence the state of the game, not the state of the server
process or computer (except through normal scenario events, such as
end of game).

For this reason, we do not load some standard lua libraries that allow
access to files or the operating system. We also disallow loading lua
libraries so that the script cannot go around this restriction.

This is the 2.2 and trunk version (Lua 5.1): we exclude the io
library, os library, package library, and blacklist functions dofile,
loadfile.

For Lua 5.1, the list of modules and functions we consider unsafe are:

  "os",
  "io",
  "package",
  "dofile",
  "loadfile",
  "module",
  "require"

These are all unavailable by not being loaded or being explicitly blocked.

See gna bug #15624

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@17157 a0f10bec-cc02-0410-94fc-a9cfff90b4cd