Make unit_info_label fixed-size as intended
Looking at the comments for get_unit_info_label_text2 in client/text.c
it is obvious that the returned sting is supposed to always contain exactly
4 lines, so that the unit info label doesn't resize itself.

However, when multiple units are selected it will instead contain
exactly 5 lines, whether the extra line is necessary or not, and thus force a
resize of the label.

This patch removes the extra line. In most cases it will only remove a
blank line. However, if 4 or more different types of units are selected
the unit info box would previously include the amount of the 3 most
common types and a summary of the rest, while with the patch it will only list
the 2 most common types and a summary of the rest.

patch by Jon Severinsson (jon_severinsson)
small code cleanup by me

see gna patch #1684



git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@17473 a0f10bec-cc02-0410-94fc-a9cfff90b4cd