define dynamic bitvector (dbv)
from bitvector.c:

There are two types of bitvectors defined in this file:
(1) bv_*  - static bitvectors; used for data which where the length is
fixed (number of players; flags for enums; ...). They are named bv_* and
the macros BV_* are defined.
(2) dbv_* - dynamic bitvectors; its size is not known a priori but
defined by the player (map known bitvectors). This bitvectors are given
as 'struct dbv' and the information can be accessed using the functions
dbv_*(). They uses the BV_* macros.

see gna #1770


git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@17929 a0f10bec-cc02-0410-94fc-a9cfff90b4cd