add check to tile_transform
This check is needed as tile could be a virtual tile used by the client
(gtk) to check which unit tasks should be shown. If a city is on the
tile and it could be transformed to a terrain with the flag
TER_NO_CITIES this would result in an assert().

see gna bug #18549


git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@20452 a0f10bec-cc02-0410-94fc-a9cfff90b4cd