Keep login network protocol compatible with former versions.
Packets headers contain 3 bytes in the initial protocol (2 for the length,
1 for the type), and 4 bytes after the SERVER_JOIN_REPLY packet has been
sent or received (2 for the length, 2 for the type).

These values are defined in "common/packets.c", respectively in functions
packet_header_init() and packet_header_set().

Label packets in "common/packets.def" which are used in the initial protocol
(their number must be in range 0-255). Update comments and documentation.

Do not send ping when accepting a new connection, because we would be unable
to handle packets with different header sizes asynchronously.

See gna bug #19943

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@22421 a0f10bec-cc02-0410-94fc-a9cfff90b4cd