Don't use four bytes to transfer a number >= 1400
The field chunk_length in the packet PLAYER_ATTRIBUTE_CHUNK can never be
bigger than 1400. It is possible to represent 1400 using two bytes. Change
it to use two bytes in stead of four.

This change makes it easier to auto generate support for the Freeciv
protocol in languages that use a signed int as the array index. It removed
the need to check that chunk_length never can have a larger value than an
unsigned int can represent.

See patch #4178

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@23412 a0f10bec-cc02-0410-94fc-a9cfff90b4cd