Show all possible actions in the GTK2, Qt, SDL and SDL2 clients.
Make the GTK2, Qt, SDL and SDL2 clients aware of the fact that the knowledge
they use to determine if an action is enabled may be uncertain. If the
knowledge used is seen as uncertain the action is shown even if it,
according to the players own (possibly wrong) knowledge, is disabled. No
marking of actions that may not be enabled, like the GTK3 client has, is
introduced.

See patch #4182

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@23421 a0f10bec-cc02-0410-94fc-a9cfff90b4cd