Move the "Unbribable" unit flag to the rule set
Make "Unbribable" a user flag in the bundled rulesets that use it. Make
"Unbribable" a not present requirement on the target of the action enabler
"Bribe Unit".

This changes the UI from displaying the bribe unit spy action and
complaining if chosen to not showing it at all. Since the type of an unit
isn't secret information this is an UI improvement.

See patch #4212

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@23485 a0f10bec-cc02-0410-94fc-a9cfff90b4cd