Add a graphics alt tag to the specialist definition
Adding an extra specialist kind to a ruleset will cause any Freeciv client
loading it to exit unless its tile set specifically support the new
specialist kind. Improve this situation by allowing the rule set author to
set an alternative graphics to be used in the specialist definition.

The alternative graphics is a normal graphics tag in the field graphic_alt.
This allows reusing small icons normally used for other things than
specialist. The confusion of having the specialist icons for different kinds
of specialist look the same can therefore be avoided. The down side is that
different images based on the citizen number won't be supported.

See patch #4239

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@23560 a0f10bec-cc02-0410-94fc-a9cfff90b4cd