Remove GTK mnemonic from two strings used in the Qt client
In the GTK clients an underscore in a label makes the following character a
mnemonic. In the Qt client it gets displayed as common text. The underscore
in the text labeling the buttons for the spy/diplomat actions "Sabotage City"
and "Establish Embassy" is therefore useless in the Qt client.

See patch #4332

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@23850 a0f10bec-cc02-0410-94fc-a9cfff90b4cd