Xaw client: Remove hard coded dependency on the Spy flag in spy actions against a city
The actions an unit with the Diplomat flag can do against a city is now
determined by action enablers, not by the presence of the Spy flag. Merging
the dialogs makes the Xaw client do the right thing in rule sets where a
diplomat without the Spy flag can to targeted actions and in rule sets where
a unit with the Spy flag can do untargeted actions.

See patch #4472

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@24347 a0f10bec-cc02-0410-94fc-a9cfff90b4cd