Use choose_action_queue_next() when the intention is to move the diplomat queue along
Calling process_diplomat_arrival() with NULL as the first parameter moves
the diplomat queue along. It is not clear from its name or its parameters
that this is what will happen.

Define a new function choose_action_queue_next() that wraps
process_diplomat_arrival(). Use it when the intention is to move the
diplomat queue along.

See patch #4514

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@24519 a0f10bec-cc02-0410-94fc-a9cfff90b4cd