Replace the No_Incite effect with the Incite City action enabler.
Rules to prevent inciting a city can be expressed using target requirements in
the action enablers of the Incite City action. The No_Incite effect is
therefore redundant. Remove it for the sake of simplicity. Port all rules in the
bundled rulesets from No_incite to Incite City.

See patch #4624

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@24705 a0f10bec-cc02-0410-94fc-a9cfff90b4cd