AI: Assume that other cities in range get action target immunity too
This will increase the motivation to build a wonder that makes many cities
immune to an action. Like the similar code for effects it assumes the result for
the city applies to the other cities in range. This is a lot faster than
checking if each city in range can be a target for an action and if the
situation will change if the evaluated building is built.

See patch #4627

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@24717 a0f10bec-cc02-0410-94fc-a9cfff90b4cd