Test nativity of moves
    While the nativity of the tiles for a move is tested, the nativity
of the actual move is not tested.  As a result, in the event that nativity
is provided by a road, the move_mode is not enforced for the provision of
nativity, possibly causing odd results, such as unit classes native to
rivers being able to cross continents.

Requested by Jacob Nevins <jtn>
Based on earlier work by Marko Lindqvist <cazfi>

See bug #16383

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@24826 a0f10bec-cc02-0410-94fc-a9cfff90b4cd