Helptext: Don't say that a unit is unbribable more than once.
Each user flag adds its description to the generated help text of all unit types
that has it. If an unit type is immune to an action (since no action enabler can
be active against it) this fact is pointed out in its auto generated help text.

Remove the user flag "Unbribable" from all bundled rulesets. Replace its use as
a !present requirement in the "Bribe Unit" action enablers with !present
requirements of the units that had it.

See bug #22255

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@25383 a0f10bec-cc02-0410-94fc-a9cfff90b4cd