Introduce a facility to find out if a unit can act given a certain unit state
The new function can_unit_act_when_ustate_is() makes it possible to find out if
a unit can perform any action controlled by generalized action enablers at all
given that a unit state property has the given value. Since going over all the
action enablers take time the values are cached.

For now this is only used to find out if a unit is of a type that can perform at
least one action from inside a transport over non native terrain. If it is true
the path finding code will allow it to go directly to its target from the
transport (without having to go to native terrain first).  It is also used to
determine if an error message should mention that the reason for the failure was
non native terrain.

See patch #5020

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@25902 a0f10bec-cc02-0410-94fc-a9cfff90b4cd