Make key_unit_diplomat_actions() dumber and therefore less redundant
Before spy acting against own tile key_unit_diplomat_actions() will check that
there is a city on it. At the moment this check isn't wrong. Spy actions vs
a unit that isn't alone at its tile are illegal. The only other target type is
cities. It could even be said to be slightly beneficial: if a client that would
ask that a unit acts against its own tile and no city is present will save a
round trip to the server.

The problem with the check is that its hard to discover. It would cause issues
if existing actions vs units are changed to no longer require the unit to alone
at its til or new actions that can target non cities are added. Remove it.

See patch #5035

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@25918 a0f10bec-cc02-0410-94fc-a9cfff90b4cd