Focus on the acting unit before asking for extra information about a spy action
The spy action targeted sabotage requires the player to select a target. The spy
actions bribe unit and incite city requires the player to give a green light to
spend the money. Focus on the actor before asking the player for feed back
so it is clear who asks.

See patch #5063

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@25996 a0f10bec-cc02-0410-94fc-a9cfff90b4cd