SDL clients: don't use client side knowledge to evaluate action enablers
Checking for what diplomatic actions are possible in the client can give the
wrong result since a rule may depend on knowledge the player don't have. The
only targets that currently are legal to act against when they share the actors
tile are foreign cities. Check for the presence of a foreign city in stead of
using possibly wrong client knowledge to check if any spy action is possible.

See bug #22510

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@26031 a0f10bec-cc02-0410-94fc-a9cfff90b4cd