Xaw client: Add information to the action selection dialog text.
If there is a target city add its name. If there is a target city and the
actor unit has a home city add the name of the home city too. The home city
name will become quite useful when the action dialog takes over trade route
establishing.

See patch #5204

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@26315 a0f10bec-cc02-0410-94fc-a9cfff90b4cd