Gtk clients: Action selection dialog: Remove the target exists checks.
Stop checking that a target of the proper kind exists before testing if each
individual action with that kind of target may be possible. It is no longer
required. The fact that the action is disabled will prevent its button from
being drawn at all.

Removing it slows down popup_action_selection() an unnoticeable amount.
Since popup_action_selection() only is called when popping up the action
selection dialog it makes no sense to keep it as an optimization. Get it out
of the way.

See patch #5287

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@26594 a0f10bec-cc02-0410-94fc-a9cfff90b4cd