unithand.c: Use client supplied action type in stead of hard coding it.
The functions unit_action_query() and unit_do_action() have switch
statements based on action id. Use the action id from action_type, the
variable that just was found to contain the specified action, in stead of
hard coding it.

This will reduce the potential for text conflicts between patches if an
action is renamed.

See patch #5344

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@26730 a0f10bec-cc02-0410-94fc-a9cfff90b4cd