Remove redundant TradeRoute flag check in base_handle_unit_establish_trade()
Whether a unit should be allowed to establish a trade route or not is now action
enabler controlled. The rule that a unit must have the TradeRoute unit type flag
to establish a trade route or enter a market place is handled there. Since
nothing in base_handle_unit_establish_trade() requires the TradeRoute unit type
flag there is no reason to look for it.

See patch #5476

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@27164 a0f10bec-cc02-0410-94fc-a9cfff90b4cd