Client: Make the sanity test in key_unit_build_wonder() stricter.
The common client code's function key_unit_build_wonder() checks that the actor
unit given to it by the client has the HelpWonder unit type flag before
forwarding the request to the server. This way the server won't be bothered by
theoretically impossible requests.

Because action enablers now control the help build wonder action it is possible
to have a unit type with the HelpWonder flag that never will be able to help a
city build a wonder. Change the test to check for the unit type's ability to do
the help build wonder action instead.

See patch #5517

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@27235 a0f10bec-cc02-0410-94fc-a9cfff90b4cd