Allow Spy units to do the untargeted version of steal tech and sabotage city.
At the moment the targeted version includes the untargeted version. This will
save a player that wants to do the untargeted version some time. Should the
targeted version stop including the untargeted version allowing them to do the
untargeted version will be required to avoid a rule change.

See patch #5545

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@27284 a0f10bec-cc02-0410-94fc-a9cfff90b4cd