SDL clients: Enable the trade route button based on ability to trade
A unit may have the TradeRoute unit flag and not be able to establish a
trade route. Another unit may not have the TradeRoute unit flag and still be
able to establish a trade route. Check for the ability to establish a trade
route in stead of the TradeRoute flag when deciding if the establish trade
route button should be shown or not.

See patch #5599

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@27400 a0f10bec-cc02-0410-94fc-a9cfff90b4cd