AI: Get a fall back trade route unit to build of one exists.
Request the best unit capable of the action Establish Trade Route in stead
of requesting the best unit with the TradeRoute unit type flag and aborting
if it can't do Establish Trade Route.

This adds support for rulesets where a unit without the TradeRoute unit type
flag is capable of the Establish Trade Route action or where a unit with the
TradeRoute flag incapable of establishing a trade route is "better" than one
that can.

See patch #5602

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@27409 a0f10bec-cc02-0410-94fc-a9cfff90b4cd