AI: A transport assigned to war isn't on a spy mission (yet).
When a transport unit is seen as a better military unit than transport unit
the AI transport code will notice and forward it to the code that handles
military units.

While forwarding it would set the unit's task to AIUNIT_ATTACK. If the unit
also could do spy actions against cities the statistics code in
dai_data_phase_begin() would later understand the unit as being on a spy
mission against a city and crash when no target city was found.

Make the transport code set the unit's task to AIUNIT_NONE in stead of
AIUNIT_ATTACK when forwarding a unit to the military unit handling code. The
other consequence of this change is that the unit will release any transport
or body guard it may have before being passed to the military unit handling
code.

Reported by Marko Lindqvist <cazfi>

See bug #23107

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@27441 a0f10bec-cc02-0410-94fc-a9cfff90b4cd