civ2civ3: make free building upkeep under communism not cumulative
The wonder A.Smith's Trading Co. and the government Communism grant a
similar bonus: free upkeep for buildings that cost 1 gold.

The wonder A.Smith's Trading Co. was supposed to be useless under Communism,
but the way it was implemented buildings with cost 2 become free of upkeep
too.

Patch by David Fernandez <bardo>

See bug #23205

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@27784 a0f10bec-cc02-0410-94fc-a9cfff90b4cd