Add an event type for reporting that a unit's order was against the rules.
Enabler controlled actions are no longer limited to spy actions. It is
therefore no longer appropriate to use E_MY_DIPLOMAT_FAILED when a unit was
ordered to do an action that turned out to be against the rules. Introduce
the new event type E_UNIT_ILLEGAL_ACTION and use it in stead.

See patch #5870

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@28337 a0f10bec-cc02-0410-94fc-a9cfff90b4cd