Make it possible to directly kill a unit from Lua.
Introduce unit_kill() in the edit module and a kill() method on the Unit
class. This saves ruleset developers from the work of hacking together a
Lua unit killer by (ab)using the parts of the API that may kill a unit. It
also serves as a reminder for Freeciv developers that units may die during a
Lua callback.

See patch #5894

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@28429 a0f10bec-cc02-0410-94fc-a9cfff90b4cd