Fix unit state legality hostile actions cache.
Patch #5975 gave it a location and cleared it on initialization but didn't write
to it when a unit state permitted a hostile action. This made it look like
hostile actions always were forbidden in every unit state. Fix it by registering
what unit states allow a hostile action.

See bug #23464

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@28763 a0f10bec-cc02-0410-94fc-a9cfff90b4cd