JSON protocol: location in packet as parameter to the dio_* functions.
Take the address of the exact location in the packet to write to / read from as
a parameter in the dio_* functions for the JSON protocol. This is required so an
element of a field array can be placed inside the array and not after it.

This is only the API change, isolated so it won't be bug hiding noise.

See patch #6017

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@28836 a0f10bec-cc02-0410-94fc-a9cfff90b4cd