Qt client: Fix found vs join city menu text.
If a selected unit is in a city city_on_tile is set to true. Initialize it
to false to make sure it only will be true in those cases.

On my system the missing initialization made the menu talk about joining a
city in stead of founding one when the selected unit was outside a city.

See bug #23557

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@28970 a0f10bec-cc02-0410-94fc-a9cfff90b4cd