classic: Set force_trade_route
It became possible to just take the one time bonus in stead of establishing
a trade route when when Establish Trade Route and Enter Marketplace was
separated. The classic ruleset is supposed to be conservative.

Take away the choice between establishing a trade route and entering the
marketplace in classic. This gives it the 2.5 rules.

See patch #6095

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@29122 a0f10bec-cc02-0410-94fc-a9cfff90b4cd