Don't let the sabotage victim unit defend it self in a dipl battle.
A diplomatic battle results in the attacker or the defender being
eliminated. The player expects the victim to be sabotaged, not eliminated.

Bribe unit, the other action with a single unit target, will only allow
units that aren't targeted engage in a defensive diplomatic battle. Make
sabotage unit consistent with that.

See patch #6103

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@29150 a0f10bec-cc02-0410-94fc-a9cfff90b4cd