AI: Warn if being a target of a new action isn't handled.
The AI uses the function action_target_neg_util() to consider how bad it is
for one of its cities to be the victim of an action. It is easy to forget it
when adding a new action that targets a city.

Make the compiler warn if a new hard coded action is added without being
listed in action_target_neg_util()'s switch statement. Being reminded to
consider actions with a city target is worth having to list actions that
don't target cities.

See patch #6106

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@29157 a0f10bec-cc02-0410-94fc-a9cfff90b4cd