View losing the diplomatic battle as getting caught
A unit that is caught while trying to perform an action may, depending on
the Casus_Belli_Caught effect, cause a diplomatic incident. This is true
when the unit is caught because a dice roll went wrong. Make it true when
it is caught because it lost a battle against a defending diplomat.

The bundled rulesets makes getting caught while trying to steal a tech cause
an incident. This patch is therefore a rule change.

See patch #6101

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@29178 a0f10bec-cc02-0410-94fc-a9cfff90b4cd