Add the new unit action "Destroy City"
"Destroy City" can be used to implement scorched earth tactics. If
destroying a foreign city using a standard unit with a reasonable cost is
permitted the ruleset should probably demand that the foreign city doesn't
have any defending units in it. It isn't hard coded. A ruleset author may
wish to have "Destroy City" as a special ability of the Leader or another
high value unit.

No way to defend against this action as long as it is enabled currently
exists. A follow up patch could make it probabilistic by allowing the
Action_Odds_Pct effect to reduce its initial odds of 100%.

The population of the target city has no chance of escaping as
refugee units or by using the migration system. A ruleset setting to give
them some means of escape and some chance of making use of it could also be
added in a follow up patch.

See patch #6248

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@29577 a0f10bec-cc02-0410-94fc-a9cfff90b4cd