alien: force actions like before action enablers
The alien ruleset allows the attacks capture units and bombard. It also
allows trade route establishing. Some units can both capture units and
bombard.

It used to be impossible to do "Enter Marketplace" when it was possible to
do "Establish Trade Route". It used to be impossible to do regular attacks
when "Capture Units" or "Bombard" was possible. It used to be impossible to
do "Bombard" when "Capture Units" was possible.

Get the old behavior back by setting force_capture_units, force_bombard and
force_trade_route to TRUE.

See patch #6277

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@29672 a0f10bec-cc02-0410-94fc-a9cfff90b4cd