Revert bug #23772 (join foreign city order)
Without a fix to bug #23772 a hypothetical player with a hypothetical
earlier 3.0 save game from a hypothetical rule set that allows joining
allied cities with a hypothetical unit ordered to go to and join an allied
city will get the order upgraded to a goto and found city. The consequence
is that the unit will go to the allied city and abort its orders since there
already is a city there. The player will then have to give the order to join
the allied city one more time.

Reports from Andreas Røsdal <andreasr> shows that the patch for bug #23772
caused segfaults for real Freeciv-web players trying to load real save games
from a real Freeciv-web ruleset. A segfault for real users is much worse
than a hypothetical user having to give an order again.

I don't have a save game to reproduce the issue.

See bug #23772

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@29756 a0f10bec-cc02-0410-94fc-a9cfff90b4cd