Make actor unit type check clearer
The function is_actor_unit() may give the impression that the check is for
the unit rather than the unit type. It is easy to ignore the unit type part
of the name is_actor_unit_type() (as opposed to unit). Replace the both with
utype_may_act_at_all()

See patch #6321

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@29788 a0f10bec-cc02-0410-94fc-a9cfff90b4cd