Let requirements check if a unit is home
Add the new UnitState property OnDomesticTile. It is true if the unit is
located withing the borders of its owner.

This can be used to limit a hypotetical "Border Patrol" unit's ability to do
the "Expel Unit" action to his owner's terrain. (A Local DiplRel requirement
would here be to the unit he tries to expel)

See patch #6323

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@29790 a0f10bec-cc02-0410-94fc-a9cfff90b4cd