Mark actions requiring player specified details
Some actions supports details beyond target. Details can be anything from
the name of the city that should be founded to the id of the tech to steal.
Freeciv can't always fill in a detail if it is missing.

Mark actions where the player has to specify additional details. Make the
information easy to access by defining the new function
action_requires_details()

See patch #6338

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@29847 a0f10bec-cc02-0410-94fc-a9cfff90b4cd