Qt client: Go to and... unit sub menu
Auto generated a "Go to and..." menu based on what the unit legally can do.
Since it's a goto it shouldn’t (by default) forbid what may become legal
when a unit reaches its target. (Example: Join the city that will be built
at the target tile next turn)

Keep the existing goto and do X orders in the old locations even if this
leads to duplication. They can be removed when the users are used to the new
locations.

Motivations:
* Allow the user to specify the order before the unit moves into harms way.
* The orders system can handle a lot more than the QT client currently
  sends.
* Make it easy to find a "Go to and..." unit order.

Putting it in an easy to discover context menu isn't in scope of this patch.
Neither is adding menu items for non action enabler controlled actions and
stuff like "nothing" (move) or "something" (action move).

See patch #6339

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@29848 a0f10bec-cc02-0410-94fc-a9cfff90b4cd