Set local building and utype for city targets
Set the target building and target utype for the target vector of any action
enabler that enables a unit to act against a city.

The initial semantic is that the target city's current production provides
the building type or unit type for local requirements. This makes the target
utype NULL when a building is being built and vice versa. This keeps the
action enabler semantic consistent for all unit vs city actions.

A possible future semantic change is to set the local building to the
specified building target for actions that specifies one while keeping the
current semantic for all other actions.

See patch #6344

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@29875 a0f10bec-cc02-0410-94fc-a9cfff90b4cd