Fix alien and civ2civ3 action enablers
The rule that only wonders can receive shields from the "Help Wonder" action
was moved to the ruleset in patch #6362. It was moved by adding the
requirement that the target city must be building a great wonder to the each
"Help Wonder" action enabler and by adding a new action enabler for cities
building a small wonder. The requirement that the target city must be
building a great wonder accidentally ended up on "Join City" in the alien
ruleset and on "Establish Trade Route" in the civ2civ3 ruleset. Move it to
the correct action enabler.

See bug #23890

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@29949 a0f10bec-cc02-0410-94fc-a9cfff90b4cd