Don't do Help Wonder via Disband Unit
Performing the Help Wonder action in handle_unit_disband() makes the
function more complex than it has to be. It is easy to forget to check if a
change done to do_unit_help_wonder() also should be done in
handle_unit_disband()'s Help Wonder. Remove it.

No hard coded rule has (intentionally) been changed. The Help Wonder action
now blocks Disband Unit. When the game disbands a unit because of upkeep it
should behave like it used to.

Don't upgrade old unit orders. An order to disband a unit in a city was
probably not intended as Help Wonder. ORDER_BUILD_WONDER and ORDER_DISBAND
were added in the same commit.

The AI won't start looking for a better alternative to its current behavior
when it must disband a unit.

See patch #6383

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@29998 a0f10bec-cc02-0410-94fc-a9cfff90b4cd