client goto: support (action) move as last order
Make send_goto_route() able to use the direction of the last move in the
unit order. Do it when the last order wants a direction. Abort when an order
requires a direction but no direction can be given.

At the moment the support for doing this is limited to the orders ORDER_MOVE
and ORDER_ACTION_MOVE. ORDER_PERFORM_ACTION can take a direction but isn't
supported yet.

See patch #6416

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@30073 a0f10bec-cc02-0410-94fc-a9cfff90b4cd