i18n: the Freeciv source files are UTF-8
The --from-code option was introduced in gettext version 0.12. Use it to
make gettext aware that the Freeciv input source files are encoded using
UTF-8. This makes it safe, as far as gettext is concerned, to use non ASCII
UTF-8 characters in translation comments and msgids.

Note that nothing is being changed for the C compiler. There is still a risk
that non ASCII characters in string literals are reencoded during the
compilation. A C string literal msgid should therefore avoid using non ASCII
UTF-8 characters.

See patch #6451

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@30157 a0f10bec-cc02-0410-94fc-a9cfff90b4cd