client: maximize profit when disbanding a unit
At the moment a more profitable way will block a less profitable way.

Temporary down side: it becomes impossible to order a unit to Disband if
Recycle Unit is illegal but the player don't have enough information to
know that it is. Will only happen in custom rulesets where Recycle Unit is
forbidden based on something the player don't know.

See patch #6536

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@30436 a0f10bec-cc02-0410-94fc-a9cfff90b4cd