Introduce the Action Auto Performer concept
An Action Auto Performer (AAP) rule makes an actor (try to) perform an
Action without being ordered to do so by the player. The only way the
player can prevent the actor from executing the order is to avoid its cause
or to make sure that the action is against the rules.

Network protocol support not included. Ruleset support not included.

Make the rules about what actions a unit will be forced to perform if his
upkeep in food, gold or shields can't be payed hard coded AAP data.

See patch #6610

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@30751 a0f10bec-cc02-0410-94fc-a9cfff90b4cd