Make payment trigger gold balance unit "selling".
Unit gold upkeep is payed at one point. At a later point, if the gold now
is negative, the game will try to "balance" the gold by removing the unit
and taking back the already payed upkeep. This is called "selling" it.

Make a unit a candidate for selling based on the upkeep that actually was
payed for it rather than what its current upkeep is.

See patch #6649

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@30863 a0f10bec-cc02-0410-94fc-a9cfff90b4cd