Plug "Found City" max 1 city info leak.
Check if the target tile of "Found City" is seen by the actor player before
checking if it has a city when asked if founding a city on it is legal.

The leak is probably not exploitable. A unit can currently only build a
city at the tile where it is located.

See patch #6714

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@31163 a0f10bec-cc02-0410-94fc-a9cfff90b4cd