Action enabler control paradrop.
Introduce the "Paradrop Unit" action enabler. Its target is the tile
the actor unit tries to land on. It can't be used in unit orders because
its maximum range (distance from actor to target) is larger than the actor
unit's tile and those next to it.

Move the rule that a unit must have the "Paratroopers" unit type flag to
the ruleset. Move the flag it self too. Reserve its unit type flag number
so the next new unit type flag can use it rather than having to renumber
the unit type user flags and unit role ids.

See patch #6949

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@31900 a0f10bec-cc02-0410-94fc-a9cfff90b4cd