Remove redundant canceled orders notification.
To notify the player that the orders for his unit were canceled is redundant
when the last order was canceled because the unit needed an action decision.
The action selection dialog will pop up and notify the player. It didn't
abort the orders for a unit on its way to some far away objective.

Reported by Marko Lindqvist <cazfi> and Jacob Nevins <jtn>

See bug #24414

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@31956 a0f10bec-cc02-0410-94fc-a9cfff90b4cd