Let ALLOW_CTRL set metamessage setting.
A connection with ALLOW_CTRL access level can set the meta server user
message via the metamessage command. It should therefore be allowed to set
it as a setting too.

A user that is able to set the meta server user message should be allowed to
view its current value too.

See patch #6967

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@31962 a0f10bec-cc02-0410-94fc-a9cfff90b4cd