rscompat: support partial req type changes.
Make it possible for rscompat_req_type_name_3_0() to handle situations where
some, but not all, instances of a requirement type has changed type. This is
useful in cases like when a base flag becomes an extra flag and if
diplomatic treaties becomes ruleset defined and moves out of DiplRel.

See patch #7056

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@32239 a0f10bec-cc02-0410-94fc-a9cfff90b4cd