Fair Islands: one resource extra per tile.
My fix for bug #24533 caused the Fair Islands map generator to set the extra
at once a resource is placed at a tile. It can later be overwritten by
another resource. The result is that the final resource is valid while those
that were overwritten by it remain as invalid resources.

Remove this accidental change.

See bug #24537

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@32268 a0f10bec-cc02-0410-94fc-a9cfff90b4cd