sandbox: require movement to expel a unit.
A unit should have some movement left to expel a foreign unit. Add it as a
requirement to the "Expel Unit" action enabler.

No current client is able to order a unit without any moves left to expel a
foreign unit. If the server receives an order to do so it will accept it
unless a rule forbids it.

See patch #7207

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@32662 a0f10bec-cc02-0410-94fc-a9cfff90b4cd