Hard code foreign target via ruleset adjustment.
The rule that certain actions must have a foreign target can be expressed in
a requirement vector. Putting it there makes it visible to code that gets
its information from requirement vectors.

Move the hard coding of the rule that certain actions must have a foreign
target to autoadjust_ruleset_data(). Add it to the default rulesets.

See patch #7248

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@32799 a0f10bec-cc02-0410-94fc-a9cfff90b4cd