Moving while able to act isn't Diplomat only.
The function unit_move_handling() can result in the server asking the client
to choose what action to perform to the destination tile. A parameter allows
the caller to specify that the unit should move even if it can act.

Traditionally the "move rather than ask what action to perform" parameter
was used to tell a Diplomat to move to an allied city instead of performing
a diplomatic action against it.

The "move rather than ask for action clarification" parameter was therefore
called move_diplomat_city. Rename it to move_do_not_act.

Another consequence of the "act vs clarify" parameter being diplomat only
was that callers that didn't expect to handle a unit capable of diplomatic
actions didn't care what value it had. Change the move_do_not_act argument
of unit_move_handling() callers so it reflects what the call intends to
achieve.

Reported by Marko Lindqvist <cazfi>

See bug #24665

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@32806 a0f10bec-cc02-0410-94fc-a9cfff90b4cd