Hard code domestic target via ruleset adjustment.
The rule that certain actions must have a domestic target can be expressed
in a requirement vector. Putting it there makes it visible to code that gets
its information from requirement vectors.

Move the hard coding of the rule that "Upgrade Unit" must have a foreign
target to autoadjust_ruleset_data(). Add it to the bundled rulesets.

See patch #7254

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@32819 a0f10bec-cc02-0410-94fc-a9cfff90b4cd