savegame2: tolerate a base becoming a pure extra.
Some extras doesn't naturally fit in as bases. Before 2.6 they still had to
be implemented as bases. It should therefore be expected that some rulesets
will drop the base part of an extra. 3.0 has already done this for Ruins.

Relax the restriction that the number of bases in a save file must be lesser
than or equal to the number of bases in the ruleset by giving an exception
to the rule if all bases in the save file are defined as extras.

See patch #7493

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@33307 a0f10bec-cc02-0410-94fc-a9cfff90b4cd