Introduce the EvacuateFirst unit type flag.
The only remaining hard coded function of the "Undisbandable" unit type flag
is to make the game try to rescue units with it before units without it when
their transport is destroyed and to use teleportation to save them.

Rename "Undisbandable" to "EvacuateFirst". Have the ruleset upgrading code
rename it.

See patch #6575

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@33401 a0f10bec-cc02-0410-94fc-a9cfff90b4cd