Gtk 3 clients: standardize unit selection icon.
Standardize on the unit selection dialog's unit icon in unit selection
items. Use it in the transport dialog and in the action selection dialog's
unit target switcher. (The last two both use the unit selects unit dialog)

This adds information like nationality, hit points and veteran status to the
transport dialog and in the action selection dialog. It therefore becomes
easier to tell the units apart.

See patch #7531

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@33436 a0f10bec-cc02-0410-94fc-a9cfff90b4cd