webperimental: introduce bombardment.
Give the Howitzer and the Battleship the ability to bombard. Set the max
range of the "Bombard" action to 3. That is further away than their vision
range. The idea is to simulate how they, with the help of a scout, can
attack targets beyond the horizon.

Bombardment doesn't replace regular attacks. Freeciv bombardment can't kill.
A Battleship is able to sink another ship.

See patch #7607

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@33630 a0f10bec-cc02-0410-94fc-a9cfff90b4cd