Support unlimited action max distance.
Allow a ruleset to specify that the entire map is a valid target of an
action rather than having to define a max_distance value large enough to
cover the biggest possible maps. This way the action will be able to cover
the entire map after the next time the map size increases.

Unlimited max distance is represented by the string "unlimited" in the
ruleset format and by the new special value ACTION_DISTANCE_UNLIMITED
internally.

See patch #7635

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@33665 a0f10bec-cc02-0410-94fc-a9cfff90b4cd